Feature implementation added for "Using Declarations"
