+++ csharp8.0/csharp8.0.sln
@@ -9,6 +9,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NonNullableReferenceTypes",
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SwitchExpressions", "SwitchExpressions\SwitchExpressions.csproj", "{D53A922D-1D63-4AB4-BAB7-B1D3842B218C}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UsingDeclaration", "UsingDeclaration\UsingDeclaration.csproj", "{B64F64C5-A4C1-4B32-9D2D-95D83A80525E}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -27,6 +29,10 @@ Global
 		{D53A922D-1D63-4AB4-BAB7-B1D3842B218C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D53A922D-1D63-4AB4-BAB7-B1D3842B218C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{D53A922D-1D63-4AB4-BAB7-B1D3842B218C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B64F64C5-A4C1-4B32-9D2D-95D83A80525E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B64F64C5-A4C1-4B32-9D2D-95D83A80525E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B64F64C5-A4C1-4B32-9D2D-95D83A80525E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B64F64C5-A4C1-4B32-9D2D-95D83A80525E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
